add and change files for csa parsl workflow
