+++ csa_params.ini
@@ -4,13 +4,14 @@ output_dir=./improve_output
 y_col_name = auc
 use_singularity = False
 hyperparameters_file = ./hyperparameters_default.json
-source_datasets = ["gCSI", "CCLE"]
+source_datasets = ["gCSI"]
 target_datasets = ["gCSI", "CCLE"]
-split = ["0", "1", "2"]
+split = ["0", "1"]
 model_name = PathDSP
 only_cross_study = False
-epochs = 100
-available_accelerators=["0","1"]
+epochs = 3
+available_accelerators=["4","5"]
+model_environment = PathDSP_env
 
 [Preprocess]
